+++ CHANGELOG.md
@@ -57,6 +57,7 @@ https://github.com/nwnxee/unified/compare/build8193.13...build8193.14
 
 ### Changed
 - Object: SetPosition() now has a toggle(default true) to update subareas if oObject is a creature, this means any traps/triggers at the new position will fire their events.
+- Object: `GetLocalVariable()` behavior has changed. See function documentation for details.
 - Profiler: Corrected a typo preventing the retrieval of certain event data (OVERHEAD_COMPENSTION_FORCE -> OVERHEAD_COMPENSATION_FORCE). Note: breaking change.
 
 ### Fixed